// src/api/asset/enums/file-upload-error.enum.ts
export enum FileUploadErrorEnum {
  AccessDenied = 'AccessDenied',
  AccountProblem = 'AccountProblem',
  AllAccessDisabled = 'AllAccessDisabled',
  AmbiguousGrantByEmailAddress = 'AmbiguousGrantByEmailAddress',
  AuthorizationHeaderMalformed = 'AuthorizationHeaderMalformed',
  BadDigest = 'BadDigest',
  BucketAlreadyExists = 'BucketAlreadyExists',
  BucketAlreadyOwnedByYou = 'BucketAlreadyOwnedByYou',
  BucketNotEmpty = 'BucketNotEmpty',
  CredentialsNotSupported = 'CredentialsNotSupported',
  CrossLocationLoggingProhibited = 'CrossLocationLoggingProhibited',
  EntityTooLarge = 'EntityTooLarge',
  EntityTooSmall = 'EntityTooSmall',
  ExpiredToken = 'ExpiredToken',
  IllegalVersioningConfigurationException = 'IllegalVersioningConfigurationException',
  IncompleteBody = 'IncompleteBody',
  IncorrectNumberOfFilesInPostRequest = 'IncorrectNumberOfFilesInPostRequest',
  InlineDataTooLarge = 'InlineDataTooLarge',
  InternalError = 'InternalError',
  InvalidAccessKeyId = 'InvalidAccessKeyId',
  InvalidAddressingHeader = 'InvalidAddressingHeader',
  InvalidArgument = 'InvalidArgument',
  InvalidBucketName = 'InvalidBucketName',
  InvalidBucketState = 'InvalidBucketState',
  InvalidDigest = 'InvalidDigest',
  InvalidEncryptionAlgorithmError = 'InvalidEncryptionAlgorithmError',
  InvalidLocationConstraint = 'InvalidLocationConstraint',
  InvalidObjectState = 'InvalidObjectState',
  InvalidPart = 'InvalidPart',
  InvalidPartOrder = 'InvalidPartOrder',
  InvalidPayer = 'InvalidPayer',
  InvalidPolicyDocument = 'InvalidPolicyDocument',
  InvalidRange = 'InvalidRange',
  InvalidRequest = 'InvalidRequest',
  InvalidSOAPRequest = 'InvalidSOAPRequest',
  InvalidSecurity = 'InvalidSecurity',
  InvalidStorageClass = 'InvalidStorageClass',
  InvalidTag = 'InvalidTag',
  InvalidTargetBucketForLogging = 'InvalidTargetBucketForLogging',
  InvalidToken = 'InvalidToken',
  InvalidURI = 'InvalidURI',
  KeyTooLongError = 'KeyTooLongError',
  MalformedACLError = 'MalformedACLError',
  MalformedPOSTRequest = 'MalformedPOSTRequest',
  MalformedXML = 'MalformedXML',
  MaxMessageLengthExceeded = 'MaxMessageLengthExceeded',
  MaxPostPreDataLengthExceededError = 'MaxPostPreDataLengthExceededError',
  MetadataTooLarge = 'MetadataTooLarge',
  MethodNotAllowed = 'MethodNotAllowed',
  MissingAttachment = 'MissingAttachment',
  MissingContentLength = 'MissingContentLength',
  MissingRequestBodyError = 'MissingRequestBodyError',
  MissingSecurityElement = 'MissingSecurityElement',
  MissingSecurityHeader = 'MissingSecurityHeader',
  NoLoggingStatusForKey = 'NoLoggingStatusForKey',
  NoSuchBucket = 'NoSuchBucket',
  NoSuchBucketPolicy = 'NoSuchBucketPolicy',
  NoSuchFile = 'NoSuchFile',
  NoSuchKey = 'NoSuchKey',
  NoSuchLifecycleConfiguration = 'NoSuchLifecycleConfiguration',
  NoSuchUpload = 'NoSuchUpload',
  NoSuchVersion = 'NoSuchVersion',
  NotImplemented = 'NotImplemented',
  NotSignedUp = 'NotSignedUp',
  OperationAborted = 'OperationAborted',
  PreconditionFailed = 'PreconditionFailed',
  RequestIsNotMultiPartContent = 'RequestIsNotMultiPartContent',
  RequestTimeTooSkewed = 'RequestTimeTooSkewed',
  RequestTimeout = 'RequestTimeout',
  RequestTorrentOfBucketError = 'RequestTorrentOfBucketError',
  RestoreAlreadyInProgress = 'RestoreAlreadyInProgress',
  ServiceUnavailable = 'ServiceUnavailable',
  SignatureDoesNotMatch = 'SignatureDoesNotMatch',
  SlowDown = 'SlowDown',
  TokenRefreshRequired = 'TokenRefreshRequired',
  TooManyBuckets = 'TooManyBuckets',
  UnexpectedContent = 'UnexpectedContent',
  UnresolvableGrantByEmailAddress = 'UnresolvableGrantByEmailAddress',
  UnsupportedMediaType = 'UnsupportedMediaType',
  UserKeyMustBeSpecified = 'UserKeyMustBeSpecified',
}